refactor: conftest
